make a repetitive action on array

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,6 +1,6 @@
 #include "function_pointers.h"
 /**
- * array_iterator - function that executes an action on each element of an arraY
+ * array_iterator - function that executes an action on each element of an a
  * @array: array to be done to
  * @size: size of the array
  * @action: action to be done
@@ -8,7 +8,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int iter;
+	size_t iter;
 
 	for (iter = 0; iter < size; iter++)
 		action(array[iter]);
